+++ applications/vessel-history/src/features/profile/Profile.tsx
@@ -27,6 +27,7 @@ import { selectDatasets } from 'features/datasets/datasets.slice'
 import { fetchResourceThunk } from 'features/resources/resources.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import { resetFilters } from 'features/event-filters/filters.slice'
+import { selectVesselDataviewMatchesCurrentVessel } from 'features/vessels/vessels.selectors'
 import Info from './components/Info'
 import Activity from './components/activity/Activity'
 import styles from './Profile.module.css'
@@ -37,12 +38,13 @@ const Profile: React.FC = (props): React.ReactElement => {
   const [lastPortVisit] = useState({ label: '', coordinates: null })
   const [lastPosition] = useState(null)
   const q = useSelector(selectQueryParam('q'))
-  const vesselProfileId = useSelector(selectVesselProfileId, shallowEqual)
-  const vesselStatus = useSelector(selectVesselsStatus, shallowEqual)
+  const vesselProfileId = useSelector(selectVesselProfileId)
+  const vesselStatus = useSelector(selectVesselsStatus)
   const loading = useMemo(() => vesselStatus === AsyncReducerStatus.LoadingItem, [vesselStatus])
-  const vessel = useSelector(selectVesselById(vesselProfileId), shallowEqual)
-  const datasets = useSelector(selectDatasets, shallowEqual)
-  const resourceQueries = useSelector(selectDataviewsResourceQueries, shallowEqual)
+  const vessel = useSelector(selectVesselById(vesselProfileId))
+  const datasets = useSelector(selectDatasets)
+  const resourceQueries = useSelector(selectDataviewsResourceQueries)
+  const vesselDataviewLoaded = useSelector(selectVesselDataviewMatchesCurrentVessel)
 
   useEffect(() => {
     const fetchVessel = async () => {
@@ -90,14 +92,12 @@ const Profile: React.FC = (props): React.ReactElement => {
   }, [dispatch, vesselProfileId, datasets])
 
   useEffect(() => {
-    if (resourceQueries && resourceQueries.length > 0) {
+    if (vesselDataviewLoaded && resourceQueries && resourceQueries.length > 0) {
       resourceQueries.forEach((resourceQuery) => {
         dispatch(fetchResourceThunk(resourceQuery))
       })
     }
-  }, [dispatch, resourceQueries])
-
-  
+  }, [dispatch, loading, resourceQueries, vessel, vesselDataviewLoaded])
 
   const tabs: Tab[] = useMemo(
     () => [