+++ applications/vessel-history/src/types/index.ts
@@ -3,7 +3,7 @@ import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 
 export type WorkspaceViewportParam = 'latitude' | 'longitude' | 'zoom'
 export type WorkspaceTimeRangeParam = 'start' | 'end'
-export type WorkspaceStateProperty = 'q' | 'dataviewInstances' | 'version' | 'vessel' | 'searchType'
+export type WorkspaceStateProperty = 'q' | 'dataviewInstances' | 'version' | 'vessel'
 export type WorkspaceAdvancedSearchParam =
   | 'IMO'
   | 'MMSI'
@@ -25,7 +25,6 @@ export type BivariateDataviews = [string, string]
 
 export type WorkspaceState = {
   q?: string
-  searchType?: string
   version?: string
   dataviewInstances?: Partial<UrlDataviewInstance[]>
   vessel?: string