+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.47
+
+### Patch Changes
+
+- Updated dependencies [16cf0830]
+  - @globalfishingwatch/react-hooks@14.0.1
+
+## 1.1.46
+
+### Patch Changes
+
+- Updated dependencies [1bb0f2b6]
+  - @globalfishingwatch/ui-components@11.1.1
+
 ## 1.1.45
 
 ### Patch Changes